Add e2e multicast testing.
Create a multicast source in one node and 2 listener in another
node.
One of the listener should join the group and receive the multicast
traffic, and the other listener should try to join another group
that does not exist and fail to receive any multicast traffic.

Organize the e2e folder correctly to avoid issues with go modules and
imports.

Disable IPv6 multicast tests for now because iperf in agnhost images
doesn't support binding a server to an IPv6 multicast group:

  bash-4.3# iperf -s -B ff00:0:3:3::3 -u -t 30 -i 5 -V
  error: Try again

Co-authored-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Antonio Ojea <aojea@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>