Reintroduce the cluster-wide multicast Port_Group.
Commit 40a90f06644a removed the multicast deny port group and used
instead the cluster port group.  However, this breaks pod-to-pod
multicast when pods reside on different nodes.  That is because OVN ACLs
are applied on all logical switch ports, including logical switch ports
connected to router ports.

Hence, an ACL of the form "if ip.mcast then deny" applied on the
clusterPortGroup will drop all multicast traffic that would normally be
routed by the cluster router even when multicast is allowed for a
namespace.

Instead, the ACL needs to match on inport/outport.

One option would be to add all POD ports to the clusterPortGroup but
that means that every time a POD is added/removed all the ACLs defined
on the clusterPortGroup (including service reject ACLs) would have to be
reprocessed by ovn-controller.  This is inefficient.

Therefore, we go back to the approach of defining a multicast specific
deny port group that includes all logical switch ports corresponding to
PODs.

Fixes: 40a90f06644a ("Migrate default deny multicast policy to port-group")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>