Support Global Availability zone.
Any node with annotation "k8s.ovn.org/ovn-az-name" set to a value
other than "global-az",  is not considered as part of Global
Availability zone.  It is instead considered as a separate
availability zone.  For such nodes ovnkube-master will not create
node logical switch and other related resources in the Northbound
database.  If the annotation is not set, then the ndoe is considered
as part of "global-az".

Signed-off-by: Numan Siddique <numans@ovn.org>