ovnkube-local: Use a default route towards mgmt port instead of source routing.
In "onkube-local" mode, we can just add a default route to get all
traffic out of the node via the mgmt port.  We cannot use the policy=src
route matching on traffic originating from the local host subnet because
that would be applied before the destination routes that move traffic
towards other AZs.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>