CI: Retest only failed E2E jobs (#2824)
* CI: Retest only failed E2E jobs

Only retest failed E2E jobs out of the matrix. Do the same for DualStack
and Upgrade tests. Test results will be stored inside github's cache and
if a sub-job was marked as completed, it is skipped and its results from
the run before are going to be displayed instead.
Also cache master and PR builds.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>