model_client: Add new mutator type for updating map values.
Insert OVSDB mutations for columns of type "map" are performed only if
the value didn't already exist in the database (RFC 7047, section 5.1).
That means that mutations that update existing values must first delete
those.  Add support for that to model_client and allow users to request
it when building operationModels.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>