Enable access to default network services from user-defined networks
Add OVS flows that redirect traffic from the masquerade
subnet to the specified services to the default gateway router.
For shared gateway mode a static route is added to redirect
the service traffic to the management port. Modified the
staticRouteCleanup to ensure that the new routes are not
getting removed.

Signed-off-by: Patryk Diak <pdiak@redhat.com>