Adds e2e test: conntrack flush after ovnkube delete
Test opens a TCP connection that simulates a GCP LB environment where
the packet is redirected via iptables to a local server on a node. Note,
in GCP the LB does not DNAT the VIP, so the packet arrives to the node
with the GCP VIP on it. In OCP, we then redirect that packet to the
local kapi server running on the node.

Once the test opens the TCP connection, it leaves it open for 2 minutes
while ovnkube-node is then deleted. Post ovn-controller starting it
should not flush the conntrack in zone 0, and the test ensures that the
conntrack entry still exists.

Recent OVN regression that prompted this E2E: https://issues.redhat.com/browse/FDP-773

NOTE: The release-1.0 backport generated conflicts because on that
branch the test didn't check IPv6.

Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 45abadba6b42558d699e1d97604c6124430344b0)
Signed-off-by: Dumitru Ceara <dceara@redhat.com>