Calculates mgmt port MAC rather than storing
Today ovnkube-controller relies on ovnkube-node to annotate the mac
address so that it can properly program the OVN mgmt port. In practice
with UDN, this means cluster manager creates the node and allocates some
subnet info to the node, both the node and the ovnkube-controller side
get updates, but ovnkube-controller fails to program the first time
because it is waiting for the node side to configure the mac address.

This patch changes the behavior of ovn-kubernetes to calculate the MAC
address for the management port from the mgmt port IP address of the
first subnet for the network. For backwards compatibility, it will first
attempt to read the node annotation, and if it no mgmt MAC exists, it
will derive it from the mgmt IP of the subnet.

Signed-off-by: Tim Rozet <trozet@redhat.com>